move SpaceInput key; remove NetCdf band type; generic implicit for SpaceInput key
