+++ spark/src/test/scala/geotrellis/spark/ingest/PyramidSpec.scala
@@ -1,16 +1,10 @@
 package geotrellis.spark.ingest
 
 import geotrellis.spark._
-import geotrellis.spark.io.accumulo._
-import geotrellis.spark.io.hadoop._
-import geotrellis.spark.io.hadoop.formats.NetCdfBand
 import geotrellis.spark.tiling._
 import geotrellis.proj4.LatLng
-import geotrellis.spark.utils.SparkUtils
 import geotrellis.raster._
 
-import org.apache.accumulo.core.client.security.tokens.PasswordToken
-import org.apache.hadoop.fs.Path
 import com.github.nscala_time.time.Imports._
 import org.scalatest._
